Add files via upload
Online Order and Delivery Management System (OODMS) is a system that provides an online platform for businesses to manage their orders and delivery services. This system is designed to streamline the entire process of ordering, tracking, and delivering products to customers. The system is user-friendly and can be easily accessed from any device with an internet connection. This report will discuss the key features, benefits, and challenges of implementing an OODMS.